add measured positioning for nav rail and nav drawer for top. center and bottom
