+++ Reply/app/src/main/java/com/example/reply/ui/ReplyListContent.kt
@@ -62,6 +62,9 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.example.reply.R
 import com.example.reply.data.Email
+import com.example.reply.ui.components.EmailDetailAppBar
+import com.example.reply.ui.components.ReplyEmailListItem
+import com.example.reply.ui.components.ReplyEmailThreadItem
 import com.example.reply.ui.components.ReplySearchBar
 import com.example.reply.ui.utils.ReplyContentType
 import com.example.reply.ui.utils.ReplyNavigationType
@@ -184,231 +187,4 @@ fun ReplyEmailDetail(
             ReplyEmailThreadItem(email = email)
         }
     }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun EmailDetailAppBar(
-    email: Email,
-    isFullScreen: Boolean,
-    onBackPressed: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(24.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        if (isFullScreen) {
-            FilledIconButton(
-                onClick = { onBackPressed.invoke() },
-                modifier = Modifier.padding(8.dp),
-                colors = IconButtonDefaults.filledIconButtonColors(
-                    containerColor = MaterialTheme.colorScheme.surface,
-                    contentColor = MaterialTheme.colorScheme.onSurface
-                )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.ArrowBack,
-                    contentDescription = stringResource(id = R.string.back_button),
-                    modifier = Modifier.size(14.dp)
-                )
-            }
-        }
-        Column(
-            modifier = Modifier.weight(1f),
-            horizontalAlignment = if (isFullScreen) Alignment.CenterHorizontally else Alignment.Start
-        ) {
-            Text(text = email.subject, style = MaterialTheme.typography.titleMedium)
-            Text(
-                modifier = Modifier.padding(top = 4.dp),
-                text = "${email.threads.size} ${stringResource(id = R.string.messages)}",
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.outline
-            )
-        }
-        IconButton(
-            onClick = { /*TODO*/ },
-        ) {
-            Icon(imageVector = Icons.Default.MoreVert, contentDescription = stringResource(id = R.string.more_options_button))
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ReplyEmailListItem(
-    email: Email,
-    isSelectable: Boolean = false,
-    isSelected: Boolean = false,
-    modifier: Modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp),
-    navigateToDetail: (Long) -> Unit
-) {
-    val semanticsModifier = if (isSelectable) modifier.semantics { selected = isSelected } else modifier
-    Card(
-        modifier = semanticsModifier.clickable { navigateToDetail.invoke(email.id) },
-        colors = CardDefaults.cardColors(
-            containerColor = if (isSelected) MaterialTheme.colorScheme.primaryContainer else MaterialTheme.colorScheme.surfaceVariant
-        )
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(20.dp)
-        ) {
-            Row(modifier = Modifier.fillMaxWidth()) {
-               ReplyProfileImage(
-                    drawableResource = email.sender.avatar,
-                    description = email.sender.fullName,
-                )
-                Column(
-                    modifier = Modifier
-                        .weight(1f)
-                        .padding(horizontal = 12.dp, vertical = 4.dp),
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = email.sender.firstName,
-                        style = MaterialTheme.typography.labelMedium
-                    )
-                    Text(
-                        text = email.createAt,
-                        style = MaterialTheme.typography.labelMedium,
-                        color = MaterialTheme.colorScheme.outline
-                    )
-                }
-                IconButton(
-                    onClick = { /*TODO*/ },
-                    modifier = Modifier
-                        .clip(CircleShape)
-                        .background(MaterialTheme.colorScheme.surface)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.StarBorder,
-                        contentDescription = "Favorite",
-                        tint = MaterialTheme.colorScheme.outline
-                    )
-                }
-            }
-
-            Text(
-                text = email.subject,
-                style = MaterialTheme.typography.bodyLarge,
-                color = MaterialTheme.colorScheme.onSurface,
-                modifier = Modifier.padding(top = 12.dp, bottom = 8.dp),
-            )
-            Text(
-                text = email.body,
-                style = MaterialTheme.typography.bodyMedium,
-                maxLines = 2,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                overflow = TextOverflow.Ellipsis
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ReplyEmailThreadItem(
-    email: Email,
-    modifier: Modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp)) {
-    Card(modifier = modifier, colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(20.dp)
-        ) {
-            Row(modifier = Modifier.fillMaxWidth()) {
-              ReplyProfileImage(
-                    drawableResource = email.sender.avatar,
-                    description = email.sender.fullName,
-                )
-                Column(
-                    modifier = Modifier
-                        .weight(1f)
-                        .padding(horizontal = 12.dp, vertical = 4.dp),
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = email.sender.firstName,
-                        style = MaterialTheme.typography.labelMedium
-                    )
-                    Text(
-                        text = "20 mins ago",
-                        style = MaterialTheme.typography.labelMedium,
-                        color = MaterialTheme.colorScheme.outline
-                    )
-                }
-                IconButton(
-                    onClick = { /*TODO*/ },
-                    modifier = Modifier
-                        .clip(CircleShape)
-                        .background(MaterialTheme.colorScheme.surface)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.StarBorder,
-                        contentDescription = "Favorite",
-                        tint = MaterialTheme.colorScheme.outline
-                    )
-                }
-            }
-
-            Text(
-                text = email.subject,
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.outline,
-                modifier = Modifier.padding(top = 12.dp, bottom = 8.dp),
-            )
-            
-            Text(
-                text = email.body,
-                style = MaterialTheme.typography.bodyLarge,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 20.dp),
-                horizontalArrangement = Arrangement.spacedBy(4.dp),
-            ) {
-                Button(
-                    onClick = { /*TODO*/ },
-                    modifier = Modifier.weight(1f),
-                    colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.inverseOnSurface)
-                ) {
-                    Text(
-                        text = stringResource(id = R.string.reply),
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-                Button(
-                    onClick = { /*TODO*/ },
-                    modifier = Modifier.weight(1f),
-                    colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.inverseOnSurface)
-                ) {
-                    Text(
-                        text = stringResource(id = R.string.reply_all),
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-        }
-    }
-}
-
-
-@Composable
-fun ReplyProfileImage(
-    drawableResource: Int,
-    description: String,
-    modifier: Modifier = Modifier.size(40.dp),
-) {
-    Image(
-        modifier = modifier.clip(CircleShape),
-        painter = painterResource(id = drawableResource),
-        contentDescription = description,
-    )
 }
