+++ Reply/app/src/main/java/com/example/reply/ui/components/ReplyProfileImage.kt
@@ -1,0 +1,24 @@
+package com.example.reply.ui.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+fun ReplyProfileImage(
+    drawableResource: Int,
+    description: String,
+    modifier: Modifier = Modifier.size(40.dp),
+) {
+    Image(
+        modifier = modifier.clip(CircleShape),
+        painter = painterResource(id = drawableResource),
+        contentDescription = description,
+    )
+}
