+++ Reply/app/src/main/java/com/example/reply/ui/theme/Theme.kt
@@ -92,7 +92,7 @@ private val replyLightColorScheme = lightColorScheme(
 
 @Composable
 fun ReplyTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(),
+    darkTheme: Boolean = false,
     dynamicColor: Boolean = true,
     content: @Composable () -> Unit
 ) {
